Move joint-plot param strategy to `utils.py`
